feat(compat): move hummingbot_network_client_adapter into hb_compat
The adapter conditionally imports hummingbot (try/except ImportError) and
lived in candles_feed/core, keeping core inside the hummingbot boundary and
requiring 4 ignore_imports crutches in the parent .importlinter. Relocate it
into candles_feed/hb_compat/ (the boundary-exempt shim package) and repoint
the sole core import plus all test patch strings. Clears ADR 0001 Group A2;
parent crutch removal is a gated follow-up.
