+++ tests/unit/test_hummingbot_integration.py
@@ -138,7 +138,9 @@ async def test_create_candles_feed_with_hummingbot(mock_hummingbot_components):
     # Set up complete patching to avoid import errors
     with (
         patch("candles_feed.integration.HUMMINGBOT_AVAILABLE", True),
-        patch("candles_feed.core.hummingbot_network_client_adapter.HUMMINGBOT_AVAILABLE", True),
+        patch(
+            "candles_feed.hb_compat.hummingbot_network_client_adapter.HUMMINGBOT_AVAILABLE", True
+        ),
         patch.dict(
             "sys.modules",
             {
@@ -186,7 +188,9 @@ async def test_candles_feed_rest_with_hummingbot(mock_hummingbot_components):
     # Set up complete patching to avoid import errors
     with (
         patch("candles_feed.integration.HUMMINGBOT_AVAILABLE", True),
-        patch("candles_feed.core.hummingbot_network_client_adapter.HUMMINGBOT_AVAILABLE", True),
+        patch(
+            "candles_feed.hb_compat.hummingbot_network_client_adapter.HUMMINGBOT_AVAILABLE", True
+        ),
         patch.dict(
             "sys.modules",
             {
@@ -216,7 +220,7 @@ async def test_candles_feed_rest_with_hummingbot(mock_hummingbot_components):
 
         # Patch the get_rest_data method of NetworkClientFactory.create_client
         with patch(
-            "candles_feed.core.hummingbot_network_client_adapter.HummingbotNetworkClient.get_rest_data",
+            "candles_feed.hb_compat.hummingbot_network_client_adapter.HummingbotNetworkClient.get_rest_data",
             new_callable=AsyncMock,
         ) as mock_get_rest_data:
             # Configure the mock to return the sample data
@@ -245,7 +249,9 @@ async def test_candles_feed_ws_with_hummingbot(mock_hummingbot_components):
     # Set up complete patching to avoid import errors
     with (
         patch("candles_feed.integration.HUMMINGBOT_AVAILABLE", True),
-        patch("candles_feed.core.hummingbot_network_client_adapter.HUMMINGBOT_AVAILABLE", True),
+        patch(
+            "candles_feed.hb_compat.hummingbot_network_client_adapter.HUMMINGBOT_AVAILABLE", True
+        ),
         patch.dict(
             "sys.modules",
             {
